@@ -56,7 +56,6 @@
     "@sentry/browser": "7.29.0",
     "@types/clone-deep": "4.0.1",
     "@types/lodash.mergewith": "^4.6.6",
-    "@types/react": "17.0.52",
     "async-limiter": "2.0.0",
     "classnames": "2.3.2",
     "clone-deep": "4.0.1",
@@ -65,21 +64,20 @@
     "eslint-config-prettier": "8.6.0",
     "eslint-plugin-prettier": "4.2.1",
     "formik": "2.2.9",
-    "framer-motion": "4.1.17",
+    "framer-motion": "7.6.4",
     "immer": "9.0.17",
     "lodash.mergewith": "^4.6.2",
     "mock-socket": "^9.0.3",
     "prettier": "2.8.2",
     "process": "0.11.10",
     "prop-types": "15.8.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-ga": "3.3.1",
     "react-hot-toast": "2.4.0",
     "react-redux": "8.0.5",
     "react-router-dom": "6.6.1",
     "react-scripts": "5.0.1",
-    "react-test-renderer": "17.0.2",
     "redux": "4.2.0",
     "redux-devtools-extension": "2.13.9",
     "redux-mock-store": "1.5.4",
@@ -90,12 +88,13 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "12.1.5",
+    "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "@types/classnames": "2.3.1",
     "@types/jest": "29.2.5",
     "@types/node": "18.11.18",
-    "@types/react-dom": "17.0.18",
+    "@types/react": "18.0.24",
+    "@types/react-dom": "18.0.8",
     "@types/redux-mock-store": "1.0.3",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jest": "27.2.1",
